+++ backend/mora/auth/saml_sso/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2021- Magenta ApS
+# SPDX-License-Identifier: MPL-2.0
+
+from .base import init_app # noqa
+from .base import check_saml_authentication # noqa
+from .base import get_session_attributes # noqa
+from .base import get_session_name_id # noqa